### 2016-12-29 #150 @whedon
[PRE REVIEW]: KraljicMatrix: An R package for implementing the Kraljic Matrix to strategically analyze a firm’s purchasing portfolio
**Submitting author:** @bradleyboehmke (<a href="http://orcid.org/0000-0002-3611-8516">Bradley Boehmke</a>)
**Repository:** <a href="https://github.com/bradleyboehmke/KraljicMatrix" target ="_blank">https://github.com/bradleyboehmke/KraljicMatrix</a>
**Version:** v0.1.0
