### 2016-12-29 #150 @arfon
@noamross - is this a good candidate for rOpenSci?
