### 2016-12-29 #150 @noamross
Thanks for checking.  We don't do statistical methods packages; our [**scope**](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) is limited to data access, extraction, manipulation and facilitating reproducible workflows.
