### 2017-01-04 #150 @arfon
👋  @bradleyboehmke - many thanks for this submission. Could you confirm that this software has a research application? Please see the 'requirements for submission' under the [author guidelines](http://joss.theoj.org/about#author_guidelines).
