### 2017-01-04 #150 @bradleyboehmke
@arfon

The most established supply chain purchasing portfolio model in academic literature and practice is Kraljic’s model ([Kraljic, 1983](https://hbr.org/1983/09/purchasing-must-become-supply-management)), now known as the Kraljic Portfolio Matrix (KPM). However, one of the primary weaknesses of the KPM has been the qualitative nature of the model, which results in a subjective method for weighting and positioning suppliers or commodities in the quadrants of the KPM. Furthermore, there are no tools that assist/automate the analytic process for analysts performing supply chain purchasing portfolio analyses.

The `KraljicMatrix` package provides a quantitative framework to objectively position purchases in the KPM matrix along with identifying which products and services to begin strategic sourcing efforts with.  This quantitative framework was based on a recent [paper](https://www.dropbox.com/s/5w7xh8ldrvodfvn/Manuscript.pdf?dl=1) I co-authored and is currently finishing its review process for publication at the Journal of Purchasing & Supply Management.  Moreover, the `KraljicMatrix` package was recently published on [CRAN](https://CRAN.R-project.org/package=KraljicMatrix) and is being implemented in the Air Force organization that performs contracting and purchasing analysis on over $10B worth of goods and services.  We are also actively advancing more perspectives of the KPM process both for academic literature and organizational practice and plan to expand the `KraljicMatrix` package’s capabilities as our science around supply chain purchasing analysis advances.

Hopefully this clarifies the research application.

v/r,
Brad
