### 2017-01-20 #150 @bradleyboehmke
@arfon

Cade Saie <cade.saie@gmail.com> is also willing to review the code.
