### 2017-02-04 #170 @cmsaie
@arfon I only have one suggestion for @bradleyboehmke.  I think when you call SAVF_score it should automatically compute the optimal rho and use it, maintain the function to call it on its own as well.
