### 2017-02-04 #170 @bradleyboehmke
No problem @arfon

[![DOI](https://zenodo.org/badge/74397607.svg)](https://zenodo.org/badge/latestdoi/74397607)

Also, @cmsaie , I like your suggestion.  I'm going to change the default in the SAVF_score function so that the optimal rho value is automatically computed and used.  I'll also allow the user to override and supply a non-optimal rho value if, for some reason, they want to.  I think this will be a good improvement to the code; it helps to streamline its functionality for the user.  Great suggestion, thanks!
